refactor(viewer): move viewport code to libs
