--- libs/insight-viewer/src/Viewport/index.ts
@@ -1,2 +1,0 @@
-export { useViewport } from './useViewport'
-export type { SetViewportAction, UseViewportParams, UseViewportReturnType } from './useViewport/type'
