+++ libs/insight-viewer/src/hooks/useViewportUpdate.ts
@@ -2,8 +2,14 @@
  * @fileoverview Update cornerstone.js viewport when Viewer's viewport prop changes.
  */
 import { useEffect } from 'react'
-import { getViewport, setViewport, getDefaultViewportForImage, CornerstoneImage } from '../utils/cornerstoneHelper'
-import { formatCornerstoneViewport, formatViewerViewport } from '../utils/common/formatViewport'
+import {
+  getViewport,
+  setViewport,
+  getDefaultViewportForImage,
+  formatCornerstoneViewport,
+  formatViewerViewport,
+  CornerstoneImage,
+} from '../utils/cornerstoneHelper'
 import { Element, Viewport, OnViewportChange } from '../types'
 
 interface Prop {
