+++ libs/insight-viewer/src/utils/cornerstoneHelper/formatViewport.ts
@@ -1,7 +1,19 @@
 import { CornerstoneViewport } from '../cornerstoneHelper/types'
 import { Viewport } from '../../types'
-import { DefaultViewport } from './const'
 import { DEFAULT_VIEWPORT_OPTIONS } from '../../const'
+
+export const DefaultViewport: Viewport = {
+  scale: 0,
+  invert: false,
+  hflip: false,
+  vflip: false,
+  rotation: 0,
+  x: 0,
+  y: 0,
+  windowWidth: 0,
+  windowCenter: 0,
+  _viewportOptions: DEFAULT_VIEWPORT_OPTIONS,
+}
 
 // Format Cornerstone's viewport to viewer's viewport prop.
 export function formatViewerViewport(cornerstoneViewport: CornerstoneViewport | undefined): Viewport {
