+++ libs/insight-viewer/src/utils/cornerstoneHelper/index.ts
@@ -1,3 +1,4 @@
 export * from './utils'
+export * from './formatViewport'
 export type { CornerstoneImage, CornerstoneViewport } from './types'
 export type { WadoImageLoaderOptions } from './setWadoImageLoader'
