+++ libs/viewport/src/index.ts
@@ -1,1 +1,2 @@
-export * from './lib/viewport'
+export { useViewport } from './useViewport'
+export type { SetViewportAction, UseViewportParams, UseViewportReturnType } from './useViewport/type'
