--- libs/viewport/src/lib/viewport.tsx
@@ -1,12 +1,0 @@
-/* eslint-disable-next-line */
-export interface ViewportProps {}
-
-export function Viewport(props: ViewportProps) {
-  return (
-    <div>
-      <h1>Welcome to Viewport!</h1>
-    </div>
-  )
-}
-
-export default Viewport
